client/requirements.txt
grpcio-health-checking==1.64.0
grpcio==1.64.0
